Split setup steps into install and Docker; draft main README
The setup includes adding the user to the docker group. This
change does not take effect until after the user logs back in.
So, split the steps into two scripts so that you can install
software and add the user to the Docker group, then logout,
and run the second series of steps that need Docker.